+++ src/documentation-helpers.ts
@@ -912,14 +912,14 @@ export const depictBody = ({
   method,
   path,
   schema,
-  mimeTypes,
+  mimeType,
   makeRef,
   composition,
   brandHandling,
   paramNames,
   description = `${method.toUpperCase()} ${path} Request body`,
 }: ReqResHandlingProps<IOSchema> & {
-  mimeTypes: ReadonlyArray<string>;
+  mimeType: string;
   paramNames: string[];
 }): RequestBodyObject => {
   const bodyDepiction = excludeExamplesFromDepiction(
@@ -940,7 +940,7 @@ export const depictBody = ({
         : bodyDepiction,
     examples: depictExamples(schema, false, paramNames),
   };
-  return { description, content: fromPairs(xprod(mimeTypes, [media])) };
+  return { description, content: { [mimeType]: media } };
 };
 
 export const depictTags = <TAG extends string>(